mappings/operational_solution_translation_test.go
package mappings

import (
	_ "embed"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestOperationalSolutionTranslation(t *testing.T) {
	translation, err := OperationalSolutionTranslation()

	assert.NoError(t, err)
	assert.NotNil(t, translation)

	tMap, err := translation.ToMap()
	assert.NoError(t, err)
	assert.NotNil(t, tMap)

}

func TestOperationalSolutionTranslationVerifyFieldsArePopulated(t *testing.T) {
	translation, err := OperationalSolutionTranslation()
	assert.NoError(t, err)
	assert.NotNil(t, translation)

	assert.NoError(t, err)
	assert.NotNil(t, translation)

	assertTranslationFields(t, translation)

}
